Move public thumbnail tests to own file
