--- shared/test/fixtures/vcr/files/metadata_cached.yml
@@ -1,164 +1,0 @@
----
-http_interactions:
-- request:
-    method: get
-    uri: https://cdo-v3-files.s3.amazonaws.com/files_test/1/1/manifest.json
-    body:
-      encoding: ASCII-8BIT
-      string: ''
-    headers:
-      Content-Length:
-      - '0'
-  response:
-    status:
-      code: 404
-      message: Not Found
-    headers:
-      Content-Type:
-      - application/xml
-      Transfer-Encoding:
-      - chunked
-      Date:
-      - Fri, 12 May 2017 22:24:27 GMT
-      Server:
-      - AmazonS3
-    body:
-      encoding: UTF-8
-      string: |-
-        <?xml version="1.0" encoding="UTF-8"?>
-        <Error><Code>NoSuchKey</Code><Message>The specified key does not exist.</Message><Key>files_test/1/1/manifest.json</Key><RequestId>602639FE8CDB00A5</RequestId><HostId>q0IciRS+Zzo7TA8kpcnFASUV4saKvicnlV4Vu2wIOt/hkRB0zxAZ6RhufQkFJu51ZWQvsqigULw=</HostId></Error>
-    http_version: 
-  recorded_at: Fri, 12 May 2017 22:24:27 GMT
-- request:
-    method: get
-    uri: https://cdo-v3-files.s3.amazonaws.com/?encoding-type=url&prefix=files_test/1/1/
-    body:
-      encoding: ASCII-8BIT
-      string: ''
-    headers:
-      Content-Length:
-      - '0'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Fri, 12 May 2017 22:24:28 GMT
-      X-Amz-Bucket-Region:
-      - us-east-1
-      Content-Type:
-      - application/xml
-      Transfer-Encoding:
-      - chunked
-      Server:
-      - AmazonS3
-    body:
-      encoding: UTF-8
-      string: |-
-        <?xml version="1.0" encoding="UTF-8"?>
-        <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/"><Name>cdo-v3-files</Name><Prefix>files_test/1/1/</Prefix><Marker></Marker><MaxKeys>1000</MaxKeys><EncodingType>url</EncodingType><IsTruncated>false</IsTruncated><Contents><Key>files_test/1/1/new_filec0cc21d843b34e9afb52.html</Key><LastModified>2017-05-12T18:44:39.000Z</LastModified><ETag>&quot;458b68094909507177f21dc1c238829d&quot;</ETag><Size>14</Size><Owner><ID>cf6bd5437eaccbf2d79d5d40694e94c727ef59eb29caa52acbc32daffeb8e9e4</ID><DisplayName>site</DisplayName></Owner><StorageClass>STANDARD</StorageClass></Contents><Contents><Key>files_test/1/1/old_fileac0a7f8c2faac49775a6.html</Key><LastModified>2017-05-12T18:44:38.000Z</LastModified><ETag>&quot;458b68094909507177f21dc1c238829d&quot;</ETag><Size>14</Size><Owner><ID>cf6bd5437eaccbf2d79d5d40694e94c727ef59eb29caa52acbc32daffeb8e9e4</ID><DisplayName>site</DisplayName></Owner><StorageClass>STANDARD</StorageClass></Contents></ListBucketResult>
-    http_version: 
-  recorded_at: Fri, 12 May 2017 22:24:27 GMT
-- request:
-    method: put
-    uri: https://cdo-v3-files.s3.amazonaws.com/files_test/1/1/.metadata/thumbnail.png
-    body:
-      encoding: ASCII-8BIT
-      string: stub-thumbnail-contents
-    headers:
-      X-Amz-Meta-Abuse-Score:
-      - '0'
-      Expect:
-      - 100-continue
-      Content-Md5:
-      - HHCzASXyiid9YZeCMBw7Jw==
-      Content-Length:
-      - '23'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Fri, 12 May 2017 22:24:29 GMT
-      X-Amz-Version-Id:
-      - m.RfpyTPJzDynZpg5MbgPp.1IYzzS7jH
-      Etag:
-      - '"1c70b30125f28a277d619782301c3b27"'
-      Content-Length:
-      - '0'
-      Server:
-      - AmazonS3
-    body:
-      encoding: UTF-8
-      string: ''
-    http_version: 
-  recorded_at: Fri, 12 May 2017 22:24:28 GMT
-- request:
-    method: get
-    uri: https://cdo-v3-files.s3.amazonaws.com/files_test/1/1/.metadata/thumbnail.png
-    body:
-      encoding: ASCII-8BIT
-      string: ''
-    headers:
-      Content-Length:
-      - '0'
-  response:
-    status:
-      code: 200
-      message: OK
-    headers:
-      Date:
-      - Fri, 12 May 2017 22:24:29 GMT
-      Last-Modified:
-      - Fri, 12 May 2017 22:24:29 GMT
-      Etag:
-      - '"1c70b30125f28a277d619782301c3b27"'
-      X-Amz-Meta-Abuse-Score:
-      - '0'
-      X-Amz-Version-Id:
-      - m.RfpyTPJzDynZpg5MbgPp.1IYzzS7jH
-      Accept-Ranges:
-      - bytes
-      Content-Type:
-      - ''
-      Content-Length:
-      - '23'
-      Server:
-      - AmazonS3
-    body:
-      encoding: UTF-8
-      string: stub-thumbnail-contents
-    http_version: 
-  recorded_at: Fri, 12 May 2017 22:24:28 GMT
-- request:
-    method: get
-    uri: https://cdo-v3-files.s3.amazonaws.com/files_test/1/1/manifest.json
-    body:
-      encoding: ASCII-8BIT
-      string: ''
-    headers:
-      Content-Length:
-      - '0'
-  response:
-    status:
-      code: 404
-      message: Not Found
-    headers:
-      Content-Type:
-      - application/xml
-      Transfer-Encoding:
-      - chunked
-      Date:
-      - Fri, 12 May 2017 22:24:27 GMT
-      Server:
-      - AmazonS3
-    body:
-      encoding: UTF-8
-      string: |-
-        <?xml version="1.0" encoding="UTF-8"?>
-        <Error><Code>NoSuchKey</Code><Message>The specified key does not exist.</Message><Key>files_test/1/1/manifest.json</Key><RequestId>7AFE965C5749BD80</RequestId><HostId>dfKib+/ZySqAnVVhcCEWBW7lFhF10sjalbdK437XLYlgvxSY1CmVlpwEBPurq/wyH/+J7pUl4T0=</HostId></Error>
-    http_version: 
-  recorded_at: Fri, 12 May 2017 22:24:28 GMT
-recorded_with: VCR 3.0.3
