+++ shared/test/fixtures/vcr/publicthumbnails/everyone_thumbnail.yml
@@ -19,16 +19,16 @@ http_interactions:
       Transfer-Encoding:
       - chunked
       Date:
-      - Thu, 12 Apr 2018 00:05:58 GMT
+      - Tue, 17 Apr 2018 18:58:36 GMT
       Server:
       - AmazonS3
     body:
       encoding: UTF-8
       string: |-
         <?xml version="1.0" encoding="UTF-8"?>
-        <Error><Code>NoSuchKey</Code><Message>The specified key does not exist.</Message><Key>files_test/1/1/manifest.json</Key><RequestId>15741C1048AA2BC9</RequestId><HostId>R38Vpf+qu+YqMNv6jQrnipphPSR4AcIKclpFSh6B79Yyf70zbDFwSY02Kz5HTjLlMCKxazB2/u0=</HostId></Error>
+        <Error><Code>NoSuchKey</Code><Message>The specified key does not exist.</Message><Key>files_test/1/1/manifest.json</Key><RequestId>031271EA4B60965C</RequestId><HostId>cMbUoDSgpsUceZMcWWU95UaBpWL2d8gPWxO+ZQVVdPcImZcvVaiuiX9BwSd4RRyhUgOa9JV5se8=</HostId></Error>
     http_version: 
-  recorded_at: Thu, 12 Apr 2018 00:05:59 GMT
+  recorded_at: Tue, 17 Apr 2018 18:58:37 GMT
 - request:
     method: get
     uri: https://cdo-v3-files.s3.amazonaws.com/?encoding-type=url&prefix=files_test/1/1/
@@ -44,7 +44,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Thu, 12 Apr 2018 00:06:02 GMT
+      - Tue, 17 Apr 2018 18:58:38 GMT
       X-Amz-Bucket-Region:
       - us-east-1
       Content-Type:
@@ -59,7 +59,7 @@ http_interactions:
         <?xml version="1.0" encoding="UTF-8"?>
         <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/"><Name>cdo-v3-files</Name><Prefix>files_test/1/1/</Prefix><Marker></Marker><MaxKeys>1000</MaxKeys><EncodingType>url</EncodingType><IsTruncated>false</IsTruncated><Contents><Key>files_test/1/1/index.html</Key><LastModified>2018-02-02T01:26:01.000Z</LastModified><ETag>&quot;7c8743c596f5f74e2cd686a2c4d562c9&quot;</ETag><Size>250</Size><Owner><ID>cf6bd5437eaccbf2d79d5d40694e94c727ef59eb29caa52acbc32daffeb8e9e4</ID><DisplayName>site</DisplayName></Owner><StorageClass>STANDARD</StorageClass></Contents><Contents><Key>files_test/1/1/new-1.css</Key><LastModified>2018-02-02T01:26:20.000Z</LastModified><ETag>&quot;d770583dcb8c55c23ad88995dbc07d9d&quot;</ETag><Size>79</Size><Owner><ID>cf6bd5437eaccbf2d79d5d40694e94c727ef59eb29caa52acbc32daffeb8e9e4</ID><DisplayName>site</DisplayName></Owner><StorageClass>STANDARD</StorageClass></Contents><Contents><Key>files_test/1/1/new-2.css</Key><LastModified>2018-02-02T01:26:20.000Z</LastModified><ETag>&quot;d770583dcb8c55c23ad88995dbc07d9d&quot;</ETag><Size>79</Size><Owner><ID>cf6bd5437eaccbf2d79d5d40694e94c727ef59eb29caa52acbc32daffeb8e9e4</ID><DisplayName>site</DisplayName></Owner><StorageClass>STANDARD</StorageClass></Contents><Contents><Key>files_test/1/1/style.css</Key><LastModified>2018-02-02T01:26:01.000Z</LastModified><ETag>&quot;76b796b04e57174c64a349a4e3f649db&quot;</ETag><Size>43</Size><Owner><ID>cf6bd5437eaccbf2d79d5d40694e94c727ef59eb29caa52acbc32daffeb8e9e4</ID><DisplayName>site</DisplayName></Owner><StorageClass>STANDARD</StorageClass></Contents></ListBucketResult>
     http_version: 
-  recorded_at: Thu, 12 Apr 2018 00:06:01 GMT
+  recorded_at: Tue, 17 Apr 2018 18:58:37 GMT
 - request:
     method: put
     uri: https://cdo-v3-files.s3.amazonaws.com/files_test/1/1/.metadata/thumbnail.png
@@ -81,9 +81,9 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Thu, 12 Apr 2018 00:06:04 GMT
+      - Tue, 17 Apr 2018 18:58:38 GMT
       X-Amz-Version-Id:
-      - Dj6L8BbrqK74iWVglCSUwcFXTnj0XVqU
+      - jI.KAgXWAyjHWpmSiB5mbAlx2.cluFXE
       Etag:
       - '"d45513eb861d6c5bf983b07f18010f55"'
       Content-Length:
@@ -94,7 +94,7 @@ http_interactions:
       encoding: UTF-8
       string: ''
     http_version: 
-  recorded_at: Thu, 12 Apr 2018 00:06:03 GMT
+  recorded_at: Tue, 17 Apr 2018 18:58:38 GMT
 - request:
     method: get
     uri: https://cdo-v3-files.s3.amazonaws.com/files_test/1/1/.metadata/thumbnail.png
@@ -110,15 +110,15 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Thu, 12 Apr 2018 00:06:06 GMT
+      - Tue, 17 Apr 2018 18:58:39 GMT
       Last-Modified:
-      - Thu, 12 Apr 2018 00:06:04 GMT
+      - Tue, 17 Apr 2018 18:58:38 GMT
       Etag:
       - '"d45513eb861d6c5bf983b07f18010f55"'
       X-Amz-Meta-Abuse-Score:
       - '0'
       X-Amz-Version-Id:
-      - Dj6L8BbrqK74iWVglCSUwcFXTnj0XVqU
+      - jI.KAgXWAyjHWpmSiB5mbAlx2.cluFXE
       Accept-Ranges:
       - bytes
       Content-Type:
@@ -131,7 +131,7 @@ http_interactions:
       encoding: UTF-8
       string: stub-everyone-thumbnail
     http_version: 
-  recorded_at: Thu, 12 Apr 2018 00:06:05 GMT
+  recorded_at: Tue, 17 Apr 2018 18:58:38 GMT
 - request:
     method: get
     uri: https://cdo-v3-files.s3.amazonaws.com/files_test/1/1/.metadata/thumbnail.png
@@ -147,15 +147,15 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Thu, 12 Apr 2018 00:06:07 GMT
+      - Tue, 17 Apr 2018 18:58:39 GMT
       Last-Modified:
-      - Thu, 12 Apr 2018 00:06:04 GMT
+      - Tue, 17 Apr 2018 18:58:38 GMT
       Etag:
       - '"d45513eb861d6c5bf983b07f18010f55"'
       X-Amz-Meta-Abuse-Score:
       - '0'
       X-Amz-Version-Id:
-      - Dj6L8BbrqK74iWVglCSUwcFXTnj0XVqU
+      - jI.KAgXWAyjHWpmSiB5mbAlx2.cluFXE
       Accept-Ranges:
       - bytes
       Content-Type:
@@ -168,7 +168,7 @@ http_interactions:
       encoding: UTF-8
       string: stub-everyone-thumbnail
     http_version: 
-  recorded_at: Thu, 12 Apr 2018 00:06:06 GMT
+  recorded_at: Tue, 17 Apr 2018 18:58:38 GMT
 - request:
     method: delete
     uri: https://cdo-v3-files.s3.amazonaws.com/files_test/1/1/.metadata/thumbnail.png
@@ -184,9 +184,9 @@ http_interactions:
       message: No Content
     headers:
       Date:
-      - Thu, 12 Apr 2018 00:06:09 GMT
+      - Tue, 17 Apr 2018 18:58:40 GMT
       X-Amz-Version-Id:
-      - j6nlJsOf8xbUX..isiQO1kgEfowiodQD
+      - UZDi7z53p62nR83Ca9Zv.UdXiRyxZ3fO
       X-Amz-Delete-Marker:
       - 'true'
       Server:
@@ -195,7 +195,7 @@ http_interactions:
       encoding: UTF-8
       string: ''
     http_version: 
-  recorded_at: Thu, 12 Apr 2018 00:06:07 GMT
+  recorded_at: Tue, 17 Apr 2018 18:58:39 GMT
 - request:
     method: get
     uri: https://cdo-v3-files.s3.amazonaws.com/files_test/1/1/manifest.json
@@ -215,14 +215,14 @@ http_interactions:
       Transfer-Encoding:
       - chunked
       Date:
-      - Thu, 12 Apr 2018 00:06:09 GMT
+      - Tue, 17 Apr 2018 18:58:46 GMT
       Server:
       - AmazonS3
     body:
       encoding: UTF-8
       string: |-
         <?xml version="1.0" encoding="UTF-8"?>
-        <Error><Code>NoSuchKey</Code><Message>The specified key does not exist.</Message><Key>files_test/1/1/manifest.json</Key><RequestId>949F1FA87CF33D97</RequestId><HostId>WeDGSL2rkcMOZWk88QS9yZkIrwOBdIXUD9U30RK/YIbWNN+vpnhWpLKxHGabv7Jufssv93+/I50=</HostId></Error>
+        <Error><Code>NoSuchKey</Code><Message>The specified key does not exist.</Message><Key>files_test/1/1/manifest.json</Key><RequestId>E31E18C27398C3D4</RequestId><HostId>Dcx5fj9HUIPkciK6+6nLbyJBWt4VqvTp1XXkTWKtujBMMWb8bEhCm+EhO+cAXxm+iPWDAzjzeUc=</HostId></Error>
     http_version: 
-  recorded_at: Thu, 12 Apr 2018 00:06:09 GMT
+  recorded_at: Tue, 17 Apr 2018 18:58:46 GMT
 recorded_with: VCR 3.0.3
